requirements.txt
django-geoposition-0.3.0
